requirements.txt
flask
flask-login
flask-sqlalchemy
astropy
bcrypt
itsdangerous
